Fix x86_64-linux-android failing tests
